requirements.txt
langchain
langchain-groq
langchain-community
langchain-core
python-telegram-bot[ext]
streamlit
supabase
python-dotenv
tavily-python
pandas
